Move try_filter_ok to separate module
